+++ objects/widget_table_record_details.py
@@ -1,6 +1,8 @@
 import plotly.express as px, dash_bootstrap_components as dbc
 from dash import Dash, dcc, html, Input, Output, State, dash_table, ALL
 
+
+id = 'table_details'
 modal_table_record = dbc.Modal(
                 [
                     dbc.ModalHeader(dbc.ModalTitle('Детализация данных о пользователях онлайн', style={'fontSize': '20px'})),
@@ -22,6 +24,25 @@ widget = [ modal_table_record,
                     style_table={'height': '715px', 'overflowY': 'auto'},
                     style_header={'backgroundColor': 'Black', 'color': 'white'},
                     style_data={'backgroundColor': 'DarkSlateGray', 'color': 'white'},
-                    id='table_details'
+                    id=id
                     ),
                  ]
+
+
+def update_table_details(df, filter_values_list, n):
+    #
+    devices = ['desktop', 'mobile']
+    countries = ['India', 'Russia', 'England', 'US', 'Japan', 'China', 'Australia', 'Canada']
+
+    filter_device = devices if filter_values_list[0] == None else [filter_values_list[0]]
+    filter_country = countries if filter_values_list[1] == None else [filter_values_list[1]]
+    filter_web_service = [filter_values_list[2]] 
+
+    df_table = df[
+        (df['web_service'].isin(filter_web_service)) &
+        (df['device'].isin(filter_device)) & 
+        (df['country'].isin(filter_country))
+        ]
+    data = df_table[['id', 'web_service', 'user_id', 'device', 'country', 'user_status', 'sign_date', 'signout_date']].to_dict('records')
+    
+    return data
